push/Student Tool Kit/requirements.txt
Flask>=2.3.2
pytesseract>=0.3.13
calamari-ocr>=1.0.0
transformers>=4.34.0
torch>=2.0.1
graphviz>=0.20.1
reportlab>=3.8.0
weasyprint>=59.0
redis>=4.7.0
rq>=1.14.0
pdf2image>=1.16.3
Pillow>=10.0.0
python-dotenv>=1.0.0
